Authenticated routes

+++ server/src/routes/memories.ts
@@ -3,8 +3,16 @@ import { prisma } from "../lib/prisma";
 import { z } from "zod";
 
 export async function memoriesRoutes(app: FastifyInstance) {
-    app.get('/memories', async () => {
+    app.addHook('preHandler', async (request) => {
+        await request.jwtVerify();
+    });
+
+    app.get('/memories', async (request) => {
+        
         const memories = await prisma.memory.findMany({
+            where: {
+                userId: request.user.sub,
+            },
             orderBy: {
                 createdAt: 'asc',
             },
@@ -19,7 +27,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
         });
     });
 
-    app.get('/memories/:id', async (request) => {
+    app.get('/memories/:id', async (request, replay) => {
         const paramsSchema = z.object({
             id: z.string().uuid(),
         });
@@ -31,6 +39,10 @@ export async function memoriesRoutes(app: FastifyInstance) {
                 id,
             },
         });
+
+        if (!memory.isPublic && memory.userId !== request.user.sub) {
+            return replay.status(401).send();
+        }
 
         return memory;
     });
@@ -46,7 +58,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
         const memory = await prisma.memory.create({
             data: {
-                userId: '2f7d8f6e-28f3-4051-b439-2ba89d30d099',
+                userId: request.user.sub,
                 content,
                 coverUrl,
                 isPublic
@@ -56,7 +68,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
         return memory;
     });
 
-    app.put('/memories/:id', async (request) => {
+    app.put('/memories/:id', async (request, replay) => {
         const paramsSchema = z.object({
             id: z.string().uuid(),
         });
@@ -71,7 +83,18 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
         const { content, coverUrl, isPublic } = bodySchema.parse(request.body);
 
-        const memory = await prisma.memory.update({
+        let memory = await prisma.memory.findUniqueOrThrow({
+            where: {
+                id,
+
+            }
+        });
+
+        if(memory.userId !== request.user.sub) {
+            return replay.status(401).send();
+        }
+
+        memory = await prisma.memory.update({
             where: {
                 id,
             },
@@ -85,12 +108,23 @@ export async function memoriesRoutes(app: FastifyInstance) {
         return memory;
     });
 
-    app.delete('/memories/:id', async (request) => {
+    app.delete('/memories/:id', async (request, replay) => {
         const paramsSchema = z.object({
             id: z.string().uuid(),
         });
 
         const { id } = paramsSchema.parse(request.params);
+
+        const memory = await prisma.memory.findUniqueOrThrow({
+            where: {
+                id,
+
+            }
+        });
+
+        if(memory.userId !== request.user.sub) {
+            return replay.status(401).send();
+        }
 
         await prisma.memory.delete({
             where: {
